code review comments fixed
