+++ src/main/java/uk/gov/hmcts/reform/prl/enums/CaseNoteDetails.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import java.time.LocalDateTime;
+
 @Data
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -14,4 +16,5 @@ public class CaseNoteDetails {
     private final String caseNote;
     private final String user;
     private final String dateAdded;
+    private final LocalDateTime dateCreated;
 }
